Refactor async/await
