+++ src/redux/reservations/reservationsSlice.js
@@ -6,10 +6,9 @@ const URL = 'https://dummyjson.com/products/?limit=5';
 export const fetchUserReservations = createAsyncThunk(
   'reservations/fetchUserReservations',
   async () => {
-    const response = await fetch(URL).then(
-      (data) => data.json(),
-    );
-    return response;
+    const response = await fetch(URL)
+    const data = await response.json()
+    return data;
   },
 );
